企业内部培训系统后端

+++ api/src/main/java/com/org/entity/MCourseType.java
@@ -11,40 +11,40 @@ import java.io.Serializable;
 /**
  * <p>
  * 课程类别字典表
- *
+
  * </p>
  *
  * @author Jie
- * @since 2022-10-23
+ * @since 2022-12-03
  */
 @Data
-@EqualsAndHashCode(callSuper = false)
-@TableName("m_course_type")
+        @EqualsAndHashCode(callSuper = false)
+        @TableName("m_course_type")
 public class MCourseType implements Serializable {
 
-    private static final long serialVersionUID = 1L;
+private static final long serialVersionUID = 1L;
 
-    @TableId("id")
-    private Integer id;
+                                        @TableId("id")
+                                            private Integer id;
 
-    /**
-     * 课程类别
-     */
-    @TableField("type")
-    private String type;
+                        /**
+         * 课程类别
+         */
+                    @TableField("type")
+                            private String type;
 
-    /**
-     * 父类别ID
-     */
-    @TableField("parent_id")
-    private Long parentId;
+                        /**
+         * 父类别ID
+         */
+                    @TableField("parent_id")
+                            private Integer parentId;
 
-    /**
-     * 类型封面图片
-     */
-    @TableField("picture")
-    private String picture;
+                        /**
+         * 类型封面图片
+         */
+                    @TableField("picture")
+                            private String picture;
 
 
-}
+        }
